refactor(toml): Move Table to the root
